Include Form526 user identifier presence mapping in user profile
Includes the mapping of required Form526 user identifier data generated by the Users::Form526UserIdentifiersStatusService in the user profile returned from the main user endpoint.

The front end will use this information to display to the user any identifiers we don't have in our system, so they can pass that information on to the Contact Center.

Adds a feature flag, form_526_required_identifiers_in_user_object, and gates the presence of this data in the profile.