Example requests for VA Forms and Appeals APIs (#14961)
* Example requests for VA Forms and Appeals APIs

* Add .bru files to Dangerfile exclusion list

These files are for the Bruno app, and are used to maintain example
requests for developers; They are not used in any application code.

* Correct bash parameter expansion in CODEOWNERS script

This mistake incorrectly categorized filenames with spaces in them

* Quote file paths in audit_service_tags PR check script

Previously this mishandled filenames with spaces in them. This change
also stops the resulting list of files from including a blank final
entry.