Implement callback url for nod email notifications.
As an initial step towards consuming VANotify status updates for an impending batch email to 3000 Veterans, we are scaffolding a callback URL that, for now, only logs what VANotify sends us. Once we validate that 1) they are able to send data to this endpoint, and 2) that the data is what we expect, then in a future PR we will build out the endpoint to persist the data we get.

Because this initial PR is only for the purpose of understanding what VANotify sends us, and will only be used in staging, it is unnecesary to write specs to accompany the changes here.