Add `travel_pay_power_switch` flag
Because we're integrating with an upstream API at the same time
the upstream API is being built, I wanted to make sure that we
could closely control whether or not the vets-api endpoint were
available.

This does not mean the `power_switch` flag will live on forever,
it should and will be retired when the feature is in prod and
correctly integrated.