Add "ping" support
Adds support for a caller to request `/travel_pay/claims/` and hit the
ping endpoint of our upstream service. This is temporary, to test
integration in our dev environment and theirs.

In the future, we will not use the ping enpoint (except maybe in
synthetic tests). Also, we will likely not maintain the dev-to-dev
integration between vets-api and our upstream. We prefer to reduce
confusion by using fake data in dev.