Dbex user identifiers status check (#14858)
* Include Form526 user identifier presence mapping in user profile

Includes the mapping of required Form526 user identifier data generated by the Users::Form526UserIdentifiersStatusService in the user profile returned from the main user endpoint.

The front end will use this information to display to the user any identifiers we don't have in our system, so they can pass that information on to the Contact Center.

Adds a feature flag, form_526_required_identifiers_in_user_object, and gates the presence of this data in the profile.

* Update User LOA JSON Schema validations for 526 identifiers

Updates four JSON schema validation files to relfect the addition of the form526_required_identifier_presence key to the users profile metadata

* Add user arg to feature flags

Forgot to pass the user arg to this feature flag. We want to allow this level of control so we could release this change progressively if we wanted to

* Satisfy outstanding Rubocop warning to use anonymous positional arguments forwarding

* Fix Flipper disable in profile test

The user argument doesn't seem to work properly when enabling and disabling Flippers in specs. Confident this is working properly in the actual codebase though.

* Address style suggestion in metadata conditional

Prefer unless statement and early return to if statment in this case

---------

Co-authored-by: Eric Boehs <eric@boehs.com>