K8s feb 28 (#15715)
* Resolve bundle-audit error (#15704)

* Change rack-cors version to fix bundle-audit error

* Add ignore flag to bundle audit

* Restore rack cors version 2.0.1

* Update bundle-audit

* Fix Gemfile.lock (#15700)

* remove x86_64-darwin-22 using bundle lock --remove-platform command

* add ruby using bundle lock --add-platform command

* Dbex user identifiers status check (#14858)

* Include Form526 user identifier presence mapping in user profile

Includes the mapping of required Form526 user identifier data generated by the Users::Form526UserIdentifiersStatusService in the user profile returned from the main user endpoint.

The front end will use this information to display to the user any identifiers we don't have in our system, so they can pass that information on to the Contact Center.

Adds a feature flag, form_526_required_identifiers_in_user_object, and gates the presence of this data in the profile.

* Update User LOA JSON Schema validations for 526 identifiers

Updates four JSON schema validation files to relfect the addition of the form526_required_identifier_presence key to the users profile metadata

* Add user arg to feature flags

Forgot to pass the user arg to this feature flag. We want to allow this level of control so we could release this change progressively if we wanted to

* Satisfy outstanding Rubocop warning to use anonymous positional arguments forwarding

* Fix Flipper disable in profile test

The user argument doesn't seem to work properly when enabling and disabling Flippers in specs. Confident this is working properly in the actual codebase though.

* Address style suggestion in metadata conditional

Prefer unless statement and early return to if statment in this case

---------

Co-authored-by: Eric Boehs <eric@boehs.com>

* Adds count of records written to log message (#15705)

* Use LOA2 instead of LOA3 for datebox timestamp logic (#15655)

* revert(vaos): removed clinic stop code logging (#15693)


va.gov-team#71272

* Filter Unwanted Errors from Retrying (Dependents Forms) (#15545)

* Add filters from exception cause

* Use sidekiq skip mechanism

* Use common method to submit backup

* Move filter handling to method

* Params for filter method

* Lint issues

* Fix icn as instance var

* Missed 686c param

* Move salvage form to top of rescue block

* Same for other job

* Remove code checks

* updated cve file in build workflow

---------

Co-authored-by: Holden Hinkle <holden@holdenhinkle.com>
Co-authored-by: Nathan Burgess <nathan.burgess28@hotmail.com>
Co-authored-by: Eric Boehs <eric@boehs.com>
Co-authored-by: gia-lexa <103389320+gia-lexa@users.noreply.github.com>
Co-authored-by: Eric Tillberg <Thrillberg@users.noreply.github.com>
Co-authored-by: AJ Magdub <ahmed.magdub@va.gov>
Co-authored-by: Tyler Fink <tfink419@gmail.com>
Co-authored-by: RachalCassity <rachal.cassity@oddball.io>