Initial service connection (#15582)
* Add `travel_pay_power_switch` flag

Because we're integrating with an upstream API at the same time
the upstream API is being built, I wanted to make sure that we
could closely control whether or not the vets-api endpoint were
available.

This does not mean the `power_switch` flag will live on forever,
it should and will be retired when the feature is in prod and
correctly integrated.

* Add blocking if power_switch disabled

Blocks requests at a module level if the feature is disabled.

* Add "ping" support

Adds support for a caller to request `/travel_pay/claims/` and hit the
ping endpoint of our upstream service. This is temporary, to test
integration in our dev environment and theirs.

In the future, we will not use the ping enpoint (except maybe in
synthetic tests). Also, we will likely not maintain the dev-to-dev
integration between vets-api and our upstream. We prefer to reduce
confusion by using fake data in dev.

* Remove unused token requests

Fixed some rubocop errors and in doing so, realized for the ping
endpoint, we don't need to pass along the VA.gov token (I think).

* Adds specs for travel pay client and controller (#15840)

* Adds specs for travel pay client and controller

* Fix linting for travel pay controller spec

---------

Co-authored-by: Athif Wulandana <athif.wulandana@athifwulandana-dyl9g7mq2f.lan>

* fix PR notes

* Fix rubocop

---------

Co-authored-by: Athif Wulandana <athif@u.northwestern.edu>
Co-authored-by: Athif Wulandana <athif.wulandana@athifwulandana-dyl9g7mq2f.lan>