API-34202 Clean up unused `AppealsApi::AppealReceivedJob` and remove feature flag (#16183)
* API-34202 Permanently switch to AppealsApi::AppealSubmittedJob

Removes the feature flag that ran this job instead of
the older AppealsApi::AppealReceivedJob. AppealsApi::AppealSubmittedJob
will now be used regardless of flags.

* API-34202 Remove defunct AppealReceivedJob specs

This has job been replaced by AppealsApi::AppealSubmittedJob, and is no
longer used. Removing both the tests and the job in one PR would change
too many lines, so this removes only the tests for now.