Implement v1 Document Upload failure mailer
Implements a new mailer, EVSS::DisabilityCompensationForm::Form526DocumentUploadFailureEmail to match the initial spec.

Filenames are obfuscated if the name is over 5 characters; the first 3 and the last 2 chars in the main filename are present, and all letters are replaced by *

Note we are using the original filename supplied to us by CarrierWave when we saved the file. Converted versions of files get converted_ appended to their filename but I believe this is only seen internally. We certainly want the veteran to see the filename of the file they

Matches the date formatting we use in Form526ConfirmationEmailJob