Add error logging in retry exhaustion handler
As with the other retry exhaustion handlers, we need to rescue any errors raised and log them so they should up in DataDog in addition to Sentry where our errors log by default. Note I did not yet write a test for this for time reasons, there are no tests written for the other handler rescue blocks elsewhere but there should be. Will return to this in the PR

Fixes other minor errors