Remove error trace logging in new mailer to protect PII
Removes the error message from the retry exhaustion logging in the document upload failure mailer and logs submission id and attachment GUID instead.

Because we're operating on email and filename information, it's possible PII could end up in the error trace and we don't want that to get into DataDog.