Add form filenames to overflow text
Adds a list of veteran-supplied files to the overflowText field so we can notify the claim adjudicator of which files they should expect in the eFolder. This is to aid in recognizing when a file fails to make it to the eFolder and we need to have the veteran submit it manually.

Moves a line in app/models/saved_claim/disability_compensation.rb where the attachments section of the form526 metadata is deleted and moved to a different key called Form526Submission::FORM_526_UPLOADS to make it easier to reference the attachments in the translation object. I don't think it makes a difference to move it down one line