Add SupportingEvidenceAttachment filenames to Form 526 overflow
Lists attached veteran-supplied files in the overflowText field of the main Form 526 metadata. This will allow a claim adjudicator to cross-reference documents that end up in the eFolder against documents the veteran originally submitted. This, along with an email notification to the veteran when a document upload fails (implemented in a separate PR), is meant to close the loop on silent failures for document uploads. It's hoped that between the veteran and the claim adjudicator any missing files can be promptly mailed in physical form for VA processing.

By default this will note the number of files and list all filenames. However, if the number of files exceeds 40, we won't list the files to avoid taking up too many pages. 40 is a number that is subject to change based on some research into pass outliers for file counts on Form submissions