VA-IIR 635: Add onboarding information to user serializer (#16592)
Add `onboarding` to the serialized output of user serializer, regardless of the feature toggle status.  If feature toggle is disabled, values will be `nil`.

Once we are ready to enable the flag to all users, formal documentation in the Swagger documentation will reflect the final shape of the altered data contract.