[Document Upload Failure] List veteran-supplied documents in Form 526 Overflow Text (#16551)
* Add SupportingEvidenceAttachment filenames to Form 526 overflow

Lists attached veteran-supplied files in the overflowText field of the main Form 526 metadata. This will allow a claim adjudicator to cross-reference documents that end up in the eFolder against documents the veteran originally submitted. This, along with an email notification to the veteran when a document upload fails (implemented in a separate PR), is meant to close the loop on silent failures for document uploads. It's hoped that between the veteran and the claim adjudicator any missing files can be promptly mailed in physical form for VA processing.

By default this will note the number of files and list all filenames. However, if the number of files exceeds 40, we won't list the files to avoid taking up too many pages. 40 is a number that is subject to change based on some research into pass outliers for file counts on Form submissions

* Revert accidental test removal when creating PR

* Revert other accidental change

* Add missing file to CODEOWNERS

* Fix accidental removal of toxic exposure line in merge

* Use file data convenience method from superclass

* Refactor Form 0781 treatment based on investigation

Upon confirming based on fixtures and code elsewhere that the form0781 key is only present and populated if the form was submitted along with the form 526 by the veteran, needed to make a couple changes to reflect this fact

* Add test for overflow notes order

The order in which we display notes in the overflow text matters, adds a test to ensure these are listed in the order approved by OCTO

* Refactor maximum overflow text handling to match new spec

We have settled on asking EVSS/Lighthouse to support up to 4000 characters in the Form526 overflowText field, so we can include a list of veteran-attached filenames if they exist.

If the total size of all notes would exceed 4000 if we included a full file list, however, we will simply include the header note describing how many files the adjudicator should expect to see show up in the eFolder

* Add StatsD metrics and logging to file list events

We will be monitoring how often these file lists are being provided and when they are so long they get truncated.

This will allow us to tweak the overflow settings if necessary, and also to know how often these large file lists are being included.

* Fix linting errors