[Document Upload Failure] Veteran-uploaded document failure notification email (#16359)
* Enqueue failed upload mailer in SubmitUploads exhaustation block

Kick off a Form526DocumentUploadFailureEmail job when the Submit Uploads job fails until it exhausts its retries

* Implement v1 Document Upload failure mailer

Implements a new mailer, EVSS::DisabilityCompensationForm::Form526DocumentUploadFailureEmail to match the initial spec.

Filenames are obfuscated if the name is over 5 characters; the first 3 and the last 2 chars in the main filename are present, and all letters are replaced by *

Note we are using the original filename supplied to us by CarrierWave when we saved the file. Converted versions of files get converted_ appended to their filename but I believe this is only seen internally. We certainly want the veteran to see the filename of the file they

Matches the date formatting we use in Form526ConfirmationEmailJob

* Refactor Form526DocumentUploadFailureEmail for encapsulation

Moving filename obfuscation to SupportingEvidenceAttachment class and date formatting to Form526 Submission class so the email doesn't know too much about how this data is structured

* Add template id and api key keys in settings.yml

Add the placeholder keys to settings.yml which will eventually hold the template id and api key for our new service, benefits_disability

* Implement obscure_filename method on SupportingEvidenceAttachment

Obfuscates the attachment's file name for use in mailers, so we don't email PII
The intent is the file should still be recognizable to the veteran who uploaded it
Follows these rules:
- Only masks filenames longer than 5 characters
- Masks letters and numbers, but preserves special characters
- Includes the file extension

* Add formatted creation time helper to Form526Submission

For the new mailer we will be using the same formating we used in the Form526ConfirmationEmailJob. Add a convenience method to the Form526Submission class so we can call it elsewhere.

* Add Flipper around new mailer behavior

Add a new Flipper switch to ensure we don't send out these emails without explicitly enabling them/can easily turn them off if the need arises.

I don't feel its necessary to scope the flipper to a user; these failures happen so infrequently a canary launch likely wouldn't cause this mailer to get sent out for a while

* Add Rails logger statement to mailer

Logs non-PII metadata on mailer so we can monitor when these emails go out and what the obscured filename the veteran sees is

* Increment statsd metric on email dispatch/fix logger message

Increments a StatsD metric when the email goes out so we can create an alert for this in DataDog.

Fixes the language in the logger message which was incorrect

* Add retry exhuastion handling

Adds a Sidekiq retry exhaustion handler to the mailer job; increments appropriate metrics and logs error info

* Add error logging in retry exhaustion handler

As with the other retry exhaustion handlers, we need to rescue any errors raised and log them so they should up in DataDog in addition to Sentry where our errors log by default. Note I did not yet write a test for this for time reasons, there are no tests written for the other handler rescue blocks elsewhere but there should be. Will return to this in the PR

Fixes other minor errors

* Add clearer doc on array transition guard

* Add new files to CODEOWNERS file

* Add missing file to CODEOWNERS to pass CI check

* Move filename Regexp matchers to constant

* Change StatsD metrics keys to match common patterns elsewhere

* Shorten logging message information

* Fix mailer job linting errors

Includes refactoring perform method to get under strict method line count linting

* Add Form526JobStatus update support

Decided it was worth continuing to use the Form526JobStatus handling in the retry handler, since one is getting created anyway.

This is helpful as a tool for investigating exhausted jobs after the fact

* Remove error trace logging in new mailer to protect PII

Removes the error message from the retry exhaustion logging in the document upload failure mailer and logs submission id and attachment GUID instead.

Because we're operating on email and filename information, it's possible PII could end up in the error trace and we don't want that to get into DataDog.

* Fix typos in documentation

* Use shorter id in test to avoid linter error on long numbers without underscores

* PR Feedback: DRY up StatsD metric constants

* PR Feedback: move submission ID lookup into method

* Add form filenames to overflow text

Adds a list of veteran-supplied files to the overflowText field so we can notify the claim adjudicator of which files they should expect in the eFolder. This is to aid in recognizing when a file fails to make it to the eFolder and we need to have the veteran submit it manually.

Moves a line in app/models/saved_claim/disability_compensation.rb where the attachments section of the form526 metadata is deleted and moved to a different key called Form526Submission::FORM_526_UPLOADS to make it easier to reference the attachments in the translation object. I don't think it makes a difference to move it down one line

* Add note for claim adjudicator when Form 0781/a is present

* Slight language tweak for documents note requested by OCTO

* Only add file list to overflow text if there are less than 41 files

Tweaks the file list logic to only display a list of files if there are less than 41, otherwise just lists how many files there are. This is to avoid adding too many pages to the overflow.

40 is an educated guess of how many files we would want to list at most, we will be following up to come up with a better figure based on historical data

* Add missing file to CODEOWNERS

* Fix linting error:

* Remove file list implementation - moved to separate PR

* PR Feedback: Add note explaining dense formatting for date

* PR feedback: rename get_ methods for more idiomatic ruby

* PR Feedback: use parsed_file_data helper in attachment superclass

* PR Feedbac: Remove uneccesary top level namespacing for Rails.logger

* PR Feedback: Consolidate variable reference in filename spec

* PR Feedback: add my team to line in CODEOWNER that was missedS

* Revert change made outside of Flipper to lessen risk of deploy

* Revert accidental Gemfile changes