Implement with_tracking block in Form526DocumentUploadFailureEmail (#16871)
This mailer job was initially implemented without properly using the rather obtuse Job Tracking module in lib/sidekiq/form526_job_status_tracker/job_tracker.rb. Adds this behavior as there were some failures in testing this mailer in staging that were throwing in the job tracker and obscuring actual errors that need to be tested.

In addition, it was required to refactor these tests to use perform_async and described_class.drain, as calling perform on these jobs doesn't appear to create a job id (jid in Sidekiq callbacks) and the Job Tracker relies on this for a lot of its logging paradigm. In the wild, these jobs will be executed with perform_async anyway.