Notify when dev/staging deploys don't succeed (#17081)
* feat: Notify when dev/staging deploys don't succeed

Closes https://github.com/department-of-veterans-affairs/va.gov-team/issues/84577

This pulls the latest status from dev and https://staging-api.va.gov/v0/status
and compares the sha to the latest in master. If the time is >30 minutes, it
messages the be-cop notifications channel. It runs every ten minutes.

Originally I had this in Atlas as I thought the logic would be easiest to
implement there as there are already models that pull the commits and deploy
status but I'd need to figure out how to get a Rails env running in a workflow
or create a Sidekiq job. Once I wrote the logic I realized it was very simple to
implement in bash.

* Populate staging summary even if dev fails