Obfuscate filenames in Form526DocumentUploadFailureEmail with Xs
Alters the filename masking to protect PII in this mailer to obfuscate with the letter 'X' instead of an asterisk. It was discovered in testing we cannot obfuscate with special characters such as an asterisk, as these filenames appear in VA Notify Mailers and their templating engine uses markdown. Therefore, special characters can be interpreted as markdown and introduce formatting issues in the mailer.