84763 handle claim status response (#17444)
* 84763 handle claim status response

* 84763 correctly parse response when an array

The Response class was using Hash #with_indifferent_access to parse the
response from claim status endpoint. But the response can be a JSON array,
which  ActiveSupport::HashWithIndifferentAccess does not like and throws an
exception for. This got caught in the rescue, which returned the string body as
is, which then incorrectly got handled in subsequent code.

Secondly, the Faraday client for Travel Claims returns the body as a string (we
don't have a JSON parsing middleware attached) whereas the specs were using
hash, which masked the issue earlier. So fixed that as well.