[API-37793] Raise a `GatewayTimeout` exception when needed (#17389)
* Raise a `GatewayTimeout` exception when needed

In Benefits Documents, if the upstream service is down, log the error
and raise a `GatewayTimeout` exception. This change handles the case in
which the upstream service responds with a non-JSON (i.e., non-hash)
object (we assume the service is down).

* Remove `res` from the exception detail

* Add Hash checking logic to `upload`

The same logic that we use in `search` applies to `upload`.

* Fix `MethodLength` error

Remove the unneeded rescue block and trim the log statement linebreaks.

* Remove `GatewayTimeout` `rescue`

We want to return an empty object to the controller, so the existing
rescue suffices.

* Update spec

Change the test to expect `/search` to return an empty object and log the
GatewayTimeout.

* Remove `res` safe navigation operator

We now check that `res` is defined and raise an exception if not.