84763 Validating claim status api response for success & failure stat… (#17387)
* 84763 Validating claim status api response for success & failure statuses

* Adding rspecs

* Moving case statements to seperate method

* Fixing few rspec failures

* Fixing rspec for new changes

* 84763 handle claim status response (#17444)

* 84763 handle claim status response

* 84763 correctly parse response when an array

The Response class was using Hash #with_indifferent_access to parse the
response from claim status endpoint. But the response can be a JSON array,
which  ActiveSupport::HashWithIndifferentAccess does not like and throws an
exception for. This got caught in the rescue, which returned the string body as
is, which then incorrectly got handled in subsequent code.

Secondly, the Faraday client for Travel Claims returns the body as a string (we
don't have a JSON parsing middleware attached) whereas the specs were using
hash, which masked the issue earlier. So fixed that as well.

* 84763 handle claim status response

* New implementation using claim status response object

* 84763 refactor claim status worker

* 84763 minor log message change

---------

Co-authored-by: kanchanasuriya <ksuriyamoorthy@kindsys.us>
Co-authored-by: Gaurav Gupta <randomsync@users.noreply.github.com>
Co-authored-by: Gaurav Gupta <ggupta@kindsys.us>