chore: Use AUTOSCALING_TARGET_VALUE env var in is_ready
This syncs the HPA scaling value with the is_ready available percentage.

When the HPA value is adjusted higher, less pods are required for the
same amount of traffic. (e.g. when it's set to 30, k8s will not desire
more pods until 30% of pods are busy).

When the is_ready threshold is adjusted higher, less traffic will be
able to hit the pod before it tells Kubernetes it's not ready for more
traffic. (e.g. when it's set to 30, it will stay "ready" until 30% of
its workers are busy)