Handle Breakers::OutageException in Post911GIBillStatusesController (#17404)
* https://github.com/department-of-veterans-affairs/va-iir/issues/828

Handle Breakers::OutageException in Post911GIBillStatusesController

The service which Lighthouse uses to request GI Bill status information for a veteran occasionally experiences
outages.  Vets API uses the Breakers gem to better handle outages by raising the OutageException class.  The
ApplicationController is configured to raise a `ServiceOutage` error in these cases.

This commit adds handling for `Breakers::OutageException` in the `Post911GIBillStatusesController`'s show method,
and to the associated service class which makes a request of Lighthouse.  When a `Breakers::OutageException` is
encountered, it is simply raised, allowing the `ApplicationController` to handle it.

The ServiceOutage class was modified to include a status code of 503 in the data passed to SerializableError.
This status code represents "Service Unavailable", which is the appropriate code to use when a service outage occurs.

Tests have been updated to cover this new exception handling logic and ensure that the correct HTTP status code
and error message are returned when a Breakers::OutageException is raised.

* update services_config.yml to allow mocking of GIBS endpoint