Change `fetch` to `presence` (#17763)
`fetch` will return `nil` if the key is defined as `nil`. We pass in a
`nil` `system_name` in `generate_upload_body`, so `fetch` will not work
as expected here. Change `fetch` to `presence` and fall back to the
default if the property is `nil` or undefined.

Also add tests to enforce the default `systemName` and `trackedItemIds`.