fix: Remove nested caching for feature toggles (#17866)
The cache key was based on the last time a feature's gate changed.
Unfortunately this won't work for features that are enabled for everyone
and then disabled as the key would not bust the cache properly.

This is because Flipper doesn't set the gate to false but DELETEs the
gate completely thus making the next `MAX(updated_at)` in line the cache
key.

I suppose this could be fixed by touching the feature_toggle when the
related gate is deleted. The code change in this PR would not need to be
done.  The performance hit here isn't enough for me to go down that path
at the moment.