Implement logging hooks for Lighthouse and EVSS upload API providers
Adds the required methods and tests for logging Lighthouse and EVSS specific metrics and logs for the different lifecycle events of the jobs that will use these API providers

We don't technically need to log retries and failures, as these are already getting logged in these upload jobs. However, since we will be monitoring the release of these new providers it will be helpful to get a high-level overview of their performance and StatsD is the most efficient way to visualize this. We do want to directly log catastrophic failures (after retries have exhausted) in the context of the new providers, to separate these error logs from errors that are getting logged for submissions that are not using the new API providers