Update SupplementalDocumentUploadProvider logging for Sidekiq error format
In the jobs we are adding new upload logging for, actual caught errors that exhaust Sidekiq retries aren't passed to the sidekiq_retries_exhausted block, so we can't pass them as an object into our new error logging handler.

Instead, we have access to the string values for the error class and the error message, because those are getting passed in the retry handlers.

Updates the logging to take these arguments instead of the error itself