Refactor Document Upload Provider for better flexibility
Makes some changes to the new 'Api Provider' infrastructure for Form 526 document uplaods to make it easier to work with in the jobs we will be migrating from EVSS uploads to Lighthouse uploads.

TLDR decoupling the provider from the document file body allows us to use the logging I built in the sidekiq_retries_exhausted block, where we don't have access to the file body as we do in the job instance itself. Also tweaks the error logging since we don't have access to an error object in that block, just the error class and message