Refactor Document Upload Api Provider for better flexibility (#18185)
* Refactor Document Upload Provider for better flexibility

Makes some changes to the new 'Api Provider' infrastructure for Form 526 document uplaods to make it easier to work with in the jobs we will be migrating from EVSS uploads to Lighthouse uploads.

TLDR decoupling the provider from the document file body allows us to use the logging I built in the sidekiq_retries_exhausted block, where we don't have access to the file body as we do in the job instance itself. Also tweaks the error logging since we don't have access to an error object in that block, just the error class and message

* Revert feature flag change

Accidentally committed the change to create a feature flag specific to the BDD instructions job instead of the generic placeholder I had in place before. This will be udpated, along with features.yml and the necessary tests, when I PR the BDD instruction job migration itself