[API-39423] Add participant/poa validation concern to POA v1 /validate  (#18433)
* Add `DependentClaimantVerification` concern to v1

Create a `DependentClaimantVerification` concern that takes a
participant id and dependent name then calls BGS to verify that the
dependent name belongs to the participant.

Add the verification to the v1 Power of Attorney `/validate` endpoint
behind a feature flag.

* Handle multiple dependents

If the participant has more than one dependent, BGS returns them under
one key: `dependent`. Change verification to account for multiple
dependents and update tests.

Add cassettes and specs to Person Web Service to capture the variable
number of dependents returned (none, one, or two).

* Tidy `normalize_name`

* Strengthen guard and rename variables

Return false if the `number_of_records` key doesnʼt exist or is zero.

Remove references to BGS in the variable names (service agnostic).

* Fix error message for participant/dependent

Change the error message to indicate which form to submit to add a
dependent.

* Add participant/poa validation concern to v1

Validate that the provided participant id and POA code combination
exists in BGS. Enable the validation in v1 POA `/validate` behind a
feature flag. Add and update tests.

* Fix Gemfile.lock

* Check only for POA code existence

We only care if the POA exists in BGS. We do not need to confirm the
participant ID matches - only that it exists.

* Change concern to service; update tests

Move the dependent claimant verification from a concern to a service.