Add POA dependent claimant validation to v2 (#18505)
Refactor the Dependent Claimant Verification Service to also allow the
claimant to provide a participant ID. We prefer the participant ID
(retrieved from MPI) over the first/last name matching.

But we only have the claimant’s participant ID in v2 (and even then, it
can be hit or miss from MPI). So fallback to the name comparison if
the claimant does not provide a participant ID. v1 should still work
with the claimant‘s first and last name.

Add the dependent claimant validation to v2.

Add and update tests to reflect the above.