Overhaul EVSS API Upload Provider logging (#18762)
Adds an extensive logging paradigm for documenting uploads via the EVSSSupplementalDocumentUploadProvider. This will track all of the important events during an upload and allow us to easily query and make dashboards in DataDog

Note since the EVSS::DocumentsService has some custom error response logic that involves throwing exceptions, we don't have an easy way of capturing an actual upload response body here. But we want to log the failure event and re-raise the exception to preserve existing behavior