[Lighthouse 526 Doc Upload Migration] Fix polling to handle UNIX timestamps in milliseconds (#18954)
* Refactor Lighthouse 526 Document Upload polling for UNIX milliseconds

The infrastructure around querying Lighthouse's Benefits Documents API POST /uploads/status endpoint to get the processing status of documents that were submitted to Lighthouse was originally implemented with the assumption Lighthouse would return standard UNIX timecodes for the start and end of processing times (i.e. in seconds since the UNIX epoch). However, in testing it was revealed these are actually returned in milliseconds since the UNIX epoch.

Change overhauls the code that processes these responses to convert the timestamps properly. Additionally cleans up the BenefitsDocuments::Form526::UpdateDocumentsStatusService spec file to make it easier to work with the complex timestamp tests

* cleanup

* more cleanup