Fix 526 Document Upload provider id logging
When submission activity is logged to the Rails logger, there is a precedent for a submission_id key to mean a Form526Submission's primary key ID; when initially implemented, these API providers for Form 526 document upload were logging the submission's VBA submitted claim ID under this key.

Corrects this to log both the submitted claim id and the Form526Submission record ID with the traditional submission_id key.