Create polled document failure handler
Creates a simple service class for handling documents we successfully submitted to Lighthouse, but found out later failed Lighthouse processing. We get this information from polling for the document status on their API uploads/status endpoint.

Implements the only current behavior we do in this situation, which is to enqueue a mailer to the veteran with next steps they can take to address the situation manually