Trigger polled Lighthouse document failure handling actions
Triggers the new callback service for Lighthouse526DocumentUpload polling records. When we mark an upload failed, calls the PolledDocumentFailureHandler to take post-failure actions.

Currently the only action implemented in the service is to trigger the appropriate failure mailer to the veteran to inform them of a manual way to work around the problem.