87622 nb disability benefits kickoff failure mailers on lighthouse polled document failed (#19242)
* Create polled document failure handler

Creates a simple service class for handling documents we successfully submitted to Lighthouse, but found out later failed Lighthouse processing. We get this information from polling for the document status on their API uploads/status endpoint.

Implements the only current behavior we do in this situation, which is to enqueue a mailer to the veteran with next steps they can take to address the situation manually

* Trigger polled Lighthouse document failure handling actions

Triggers the new callback service for Lighthouse526DocumentUpload polling records. When we mark an upload failed, calls the PolledDocumentFailureHandler to take post-failure actions.

Currently the only action implemented in the service is to trigger the appropriate failure mailer to the veteran to inform them of a manual way to work around the problem.

* Fix linting errors

* Fix more linting errors