Do not swallow errors in `send_to_lighthouse!` (ZSF) (#20371)
* do not swallow errors in send_to_lighthouse! (ZSF)

When we attempt to submit a claim, we first try to send it to VBMS.  If
that action fails, we fall back to Lighthouse.  Previously the
`send_to_lighthouse!` function swallowed any exceptions it encountered.
This in turn would not trigger a retry (and eventually a failure email
message to the user) in the Sidekiq queue.  As a result, this caused a
silent failure.