Adds tests, rest of functionality
Note that because VistA and BTSSS do NOT store times in UTC, and that
some integrations (VPG, VAOS, etc) convert those times to UTC, we have
to do some translation in order to get everything to talk nicely with
each other. Essentially, we're using local time and stripping time
zone.