Update to Rack 3.0.12 (#20834)
* Update to Rack 3.0.11

* fix: Add Rack Middleware to rewrite legacy query params

Controllers using the `Filterable` concern were sending a non-compliant
URL syntax for nested filter params (e.g.
`filter[[status][eq]]=Active,Expired` instead of `filter[status][eq]=Active,Expired`). This was due to a
requirement defined in the Filterable concern in 2016.

- Add `LegacyFilterQueryRewriteMiddleware` to convert legacy filter queries
- Update `valid_filters?` regex in `Filterable` to match RFC-compliant syntax

This fix allows for both syntax to be submitted. When the invalid syntax
is submitted, the middleware corrects it. When the correct syntax is
submitted, it passes through. The new Filterable regex validates the new
syntax and Rack 3 correctly parses it. This is not backwards compatible
with Rack 2 and will need to be deployed along side the Rack 3 upgrade.

Since this update supports both syntaxes, the front end and mobile apps
will not require an update.

For more info, see: https://github.com/rack/rack/blob/main/UPGRADE-GUIDE.md#invalid-nested-query-parsing-syntax