POA decline email (#20918)
* [WIP]

* Schema changes

* [WIP] email job

* [WIP] foreign key fix

* Working migration for notifications

* rename type to avoid convention mismatch

* Redo association

* Make va notify notification optional

* Fix linting error

A new rubocop rule has been introduced recently that requires hash
key-values where the key and the value match like:a

notification_type: notification_type,

to be condensed to:

notification_type:,

* [WIP] email job spec

* Fix tests for POA email job

* [WIP] tests for notification class

* Fix POA request notification tests

* Remove unused traits

* notification_type -> type

* Remove migration

* Add decline email template_id to settings

* Email job now accepts notification and personalisation

* Make email job take existing notification record

We're going to start creating the POA request notification record before
enqueing the email job.  That way we can track when it doesn't have a
notification_id that ties it to a VANotify::Notification record.

* Enqueue decline email within decisions controller

* Add tests for sending decline emails

* Add POA request notification tracking to digital submit confirmation email

* Create notification record after guard

* Don't send email address as a sidekiq arg

We can look it up from the POA request notification record.

* Update tests